First (New) Version!
MenuDrop! It originally started from the menu system I made for Smart Text Editor, which has been around for probably over a year now. That menu system started to get fairly dated and buggy as Smart Text Editor moved on, so I thought I would rework it into it's own project using the Web Components API! I think it will be nice to be able to use it in other projects as well.
Over these next updates, I start to rewrite the system from the ground up, hopefully solving a lot of the issues that the original system had. Eventually MenuDrop will be replacing the old system used in Smart Text Editor.

When I first started the rewrite, I thought it may be an interesting idea to define the menus using JSON, and have the component create all of the HTML output. Eventually I decide to just use HTML. Why reinvent the wheel when there are already nice tools out there for you to use? (In this case, HTML haha)