Default Styles Additions
Adjusted many styles that will come default when using a MenuDrop element! I mostly reorganized a lot of the selectors to make more sense, and added CSS variables to parts of the rules that were getting repetitive. I decided to make the default styles a little less fancy as well. Funny choice to make, but it's just supposed to be good enough for the menus to work, you can add your own styles to make something great!

I also fixed a few bugs!
When you click on a divider element, the parent list will now receive focus. Also, the parent list's positioning will now be accurate to it's location on the page, even if the Visual Viewport has a higher or lower zoom level. Positioning the parent list now takes the Visual Viewport's offset values in to play to counteract this issue.

One more addition, I almost forgot! The sub-list option icon is now shaped using a CSS mask, rather than the border sizing trick to make triangles. I found that the border trick didn't work so well with text size-relative scaling, so I made an SVG mask that looks similar to the original shape, and it will instead mask a baseline square shape into the desired arrow-looking one! Pretty neat stuff, CSS is great :)