Setting Up for TypeScript && ESM
Starting to move the library up to my new workflow setup. Can't wait!
Wasn't sure what the npm version should be, as there haven't really been official releases yet. I think this will essentially be the third rework though, so I'm going with that. The first version was the one built into Smart Text Editor, then the move to the library's own repo (this current version, essentially. But that one was never really completed), and this new rework will be the third version after those. I want to complete more of what I wanted to do in the last version, as I didn't totally complete everything I wanted to do for it. Everything else I learned since working on the latest version is going to help make things a ton better. Can't wait!