TS Time!
Moving over to TypeScript! JSDoc will do, but TypeScript has been so nice, so I'm gonna try it with this component and see how much I like it :)

Lots of new testing to come, too! This component is a bit behind from the new things I've started trying with Num-Text, and I'm gonna bring those, and many more new ones, over to this next!