Reversing ElementTemplate
It was a really cool workaround! With using ES Modules instead of sync module loading, element loading times aren't a problem for accessing the `.innerHTML` anymore, so I don't have to use the workaround anymore. Now that I have experience with async code, it makes more sense of how you can expect things to not quite be ready to operate on yet. Waiting for the Web Component to populate it's content is fairly straightforward now :)

ESM for the win! You get so many benefits from it, and you learn so much because of it :D

TypeScript makes it that much cooler...