Fully Type Safe!
Finished all of the type checking to make the current implementation fully type-safe! Most of the errors had to do with variable type inferences, and nullish/undefined runtime checks. This is helping out so much! This was the biggest challenge that made it hard for the component to handle all of the unexpected instances where some elements may not be present at a given moment. I already said it, but this is helping out so much!

Coming up next: Reworks and additions!

Watched the Elvis movie tonight, that was really interesting! I feel bad for how much he had to go through, being stuck in his career for a really long time.

Listening to Lunatic Soul again, while writing this :)