Element Matches Type Guards
I wish that `Element.matches()` would perform a type guard in an if statement for the element you are checking. It would be great if it worked similar to `Element.closest()`, where you can pass the element class as a type parameter if the selector is too complicated for TSC to deduce from the `closest()` string. The first workaround I used was to assert the value to the correct type after the type guard is performed manually, but this tends to get messy if you use that variable a lot. This time I addded a second full type guard that ensures that the element is an `instanceof` of that element type you are checking. So something like this: `if (event.target.matches("button") && event.target instanceof HTMLButtonElement){ event.target; // HTMLButtonElement }`. Ideally it would be awesome if `Element.matches()` could work the same way as that above, just embedding that all into one call, maybe with the `Element.matches<T>()` syntax, then you could also use a complex selector like `button:hover:active:focus-visible`, or something weird haha.