Test Page Update
Removed some extra code from the test page. Gonna try starting fresh, now that I'm in TS land? Not sure how it's gonna go, or if I want to redo everything. But, TS has shown to make it not as hard when things are typed out in front of view. Nice and easy to ensure that things are type safe, and easy to find dumb errors first :)