Default Module Exports
Not sure if I'm gonna keep this one, but I'm gonna try it out. It seems like a lot of component-based code uses default exports for managing it's codebase with modules, and I thought it may work nicely here too. Haven't used it like this on a bigger scale yet (or, just in a while in general), so I'm gonna see how I like it :)