Constructable Stylesheets!
Oh yeah! Nice to have this working now. Rather than loading the styles for the components with a stylesheet linked by the user, now it's included as part of the library's source!

I really like this, as it makes it s the only thing required to fully load the component is to run the script from, wherever essentially, and just by it's URL. You don't have to add any other tags to the head of the page, you don't have to add any polyfills yourself, it just runs.

The only thing that could be debated to be a bit of a drawback is the FOUC, Flash of Unstyled Content, but since everything is lazy-loaded *consistently*, and with the smart things that come in handy thanks to the Custom Elements API, you can rely on different ways to handle that yourself.

I think one way to do that would be with the `window.customElements.whenDefined()` function call, which returns a `Promise` when the given element tag you are checking is defined. This allows you to get a direct callback in your app once 'x' component has fully finished loading. If you do this with all of your components' `Promise`s, then you know exactly when it's safe to fully render the page!

Using this in conjunction with a main `<script type="module"></script>` for your app, then you can simply `await` all of these calls, in the top level, and once they have finished loading, continue the main module process! Of course, if you didn't need to wait for them to finish, you can simply do this for your rendering/display logic, and hide elements as they are loading, accordingly.

I really like this kind of loading idea, where you expect things to not be ready yet. If you do that, then you can prepare for more things on the fly, and it allows it to handle a range of circumstances.

I'm going to try using this design in my web apps from here onwards, I think it can really make your code go a long way :)

I want to go back and try and bring all of my modern code to STE, but I don't know where to start every time I go over there XD. Gonna look into it more, I think ideas like these could really help with how STE is structured.

I definitely want to combine STE with NBTify down the road, that's been one of my goals for a long time now! To get to that point though, I need to step up STE's codebase a bit, and also add SNBT support to NBTify, as I think that's how I want to go about editing NBT. I think the NBT tree view editors with buttons and such look really cool, and they are fairly easy to use, but I think there's something neat about just opening it up as a text file, typing a few things, then saving that with just a keypress. If I made that into it's own mini PWA, I think the Minecraft community would flip their lids! It's definitely something I'd love to have, especially with it working offline, on mobile, and all of the other cool things you get with PWAs. I may have to start a little separate repo for that soon! Maybe it should be it's own thing, rather than being part of STE? I'd probably have both eventually, as I think a lightweight editor specifically for NBT would be nice, and just having it in STE alongside everything else seems really cool. We'll have to see! Lots of cool things ahead :)

Been looking into React and Lit recently too! With all of this fancy TypeScript and module work I've been doing for the last few months now, I feel more equipped to tackle things like that! Next I want to look into how you can get a minimal React/TypeScript setup going, and still be able to publish to GitHub Pages.

So many cool things!!! Gotta go to bed, cya ta' mornin