Legacy Types
Started working on this yesterday afternoon; git says 16 hours ago :)

This provides TypeScript types for the old version of Menu Drop which I am currently still using in the stable version of Smart Text Editor. I'm reworking STE to be more strongly typed, so having type definitions for my own libraries that I'm using is the final step to having a fully-typed out codebase!

The modern version of Menu Drop that I have been working on hasn't been UX-synonymous with this original version, so having this will also be very helpful to adjust the in-progress rework of Menu Drop so that it can be UX-compliant with this one too. Some examples include how it handles user input, key press detection, tab navigation, things like that. Since this version was semi-spaghetti code-ey, I essentially started fresh from the ground up when reworking Menu Drop, but that caused some features to fall behind, and not behave like they used to. To be able to use the new version in the modern STE, I want it to work just as well!