Type Guard Safety Checks
Encountered a specific occurrence where the new `Element.matches()` typings don't behave as expected, and turn safe values into the `never` type. I'm tracking this issue in another project that relies on a fix for this, so you can see a demo and explainer for it, being tracked here:

https://github.com/BenjaminAster/Better-TypeScript/issues/1

Also did some general code formatting and moving, now that I've discovered some more new things the last few weeks. Now that my Web Components are run using ES Module syntax, I can depend that they will be run only after the document has been loaded, which makes initializing the component in the constructor a safe option now.

Going to start leaving the end of file new line at the end of my `package-lock.json` files, as it's been frustrating to remove it every time after a simple `npm install`. With time, I think I'll finally convince myself to just leave it in all my files. Going to get there eventually, as I am getting to be more and more of a Linux guy, and that's more of an expected standard over there (In Unix-land).