Escape Key Scoping
Figured out the issue with the Escape key UX which would unexpectedly close the menu because the list-closing focus change couldn't locate a list opener properly. I had to check the old code to find why it wasn't working. Turns out it was because my `querySelector()` calls to get the target list opener wasn't like it was originally. Now it can find the correct opener again!