Menu Opener Event Target
Oooh! Realized how to make the `<menu-opener>` component the selected target for `Event.target` use.

I realized that I can wrap the element's child content inside of a Shadow Root `<slot>`, and that will make the main `<menu-opener>` element the shown element for `Event.target` use. This means you also no longer have to wrap your opener text inside of a `<button>`, since that will be part of the Shadow Root itself, and your text will be instead rendered inside of that.

This does make the opener a bit different to style now though, as next I'll make it so you can select the `<button>` in the Shadow Root using the `menu-opener::part(button)` selector. This is similar to the behavior for `input[type="file"]::file-selector-button`.

I also added a method to `MenuOpener` which will focus the internal `<button>` for you, just by calling `MenuOpener.focus()`.