Code Prettifying + Missing Event Target
MTV Unplugged for this commit, Animosity for the last commit :)

Now that the code is less buggy and also type-safe I did some formatting changes to make things a bit more easy on the eyes, as it was a little hard to read some of the event logic before.

Realized I forgot to update one more `Event.target` rework in the last commit, so I updated it in this one.

Found a neat way of working with `window.requestAnimationFrame()`, with Promises! If you only need to rely on it for calling an action, you can simply await it from a wrapped Promise, and make your event handler an async function. Not a big change for the one `focusout` event that does this, but I think it could be a very cool method for making a game loop stack a bit better, like in Flatlands. You can remove callback heck (XD) even for places like that, which I really like. I hadn't thought of using `window.requestAnimationFrame()` like that before, so it was another Promises-aha moment, which was similar to turning `HTMLImageElement.onload` events into Promises too, which is also very nice to work with. I still have to add those to Flatlands too.

https://stackoverflow.com/questions/60566546/is-it-safe-to-promisify-requestanimationframe