npm Setup
This component is going to follow a similar migration path to what I just did with Num-Text, which is getting it ready to be able to run by itself directly from npm! It will no longer need to be used from GitHub Pages anymore. Doing this involves making a bit better type definitions with some more description documentation (not just types), and formatting things a bit so I can use it to build the next generation of the component with this being a guideline for already-implemented features.