Element.matches() Type Guards
Bringing this tech over from the dev branch of Menu-Drop.

This helps remove extra type assertions that were only there because the type inferences weren't strong enough for the `Element.matches()` checks that I use throughout the codebase.

I had to change some structuring of the existing logic so it can more strongly be represented with the types. This won't change any of the existing functionality, be it buggy, or not. This stage isn't about fixing the issues with the old version, I just want it to be type safe and ready so it can be a solid module to rely on and use as a reference for the new reworks down the line.