Package Config + Module Resolution
After doing some testing with the types for the app, I am glad that I noticed that adding `moduleResolution: NodeNext` actually forces all of the exported JavaScript files to *all* be ESM, not matter if there are `import` statements present or not. This breaks the current band-aid loading for the library as it is right now, because it currently *isn't able to* support ESM yet, because STE is currently relying on it being a regular script tag, without any exports. It relies on the side effects of adding to the global scope of the window. After digging into finding out that's what caused that, simply removing `moduleResolution` reset it back to the default, and now it has the correct expected behavior again.

I do find this to be a great fix, I didn't realize you could specify a single module type like that, I wonder if I can do that with Service Workers, but the other way around. I want to treat a Service Worker like a module at the type level, but I want it to export as a regular script, without any import/export statements. Writing that out though, it actually makes sense that you can't cross those lines hazily, I'm glad that you have to be explicit one way or another with the module file. I do wish TypeScript would better handle the scope provided by individual Workers in general though. The only way to specify a file is in a different scope to another context, is to make it a module. This now enforces an `export {}` call at the end of the file though, if the module-ness isn't established somewhere else in the file already. This prevents you from getting the Worker scoping though, and you can't use Module Workers in all browsers yet unfortunately. That would be the best fix, for Module Workers to be present everywhere.

Ok, back to the changes. This updates also updates the config for the package, now it's much closer to being ready for an initial npm build. I also changed the small about text for the library, and changed the repo URL ahead of time, since I will be updating that to the new name soon, as well.