API Docs (mostly)
Added documentation to go along with the component's types. Because some of the APIs I made are both unfinished and a bit sketchy, I'm going to refrain from documenting those ones out fully, as I want to make it from scratch when I bring them up to the component rework.

I had some more stream of conscious writing, this time for inside the documentation. I thought it would be noteworthy to add to the commit itself, also. This was first about the 'select' mode, and also some was about the 'shortcuts' feature.

```
I never fully finished this functionality, so there's not much of a great API to work with this unfortunately.
Same with the 'shortcuts' feature. It works great for what this legacy version needed to do, but it really could
use a fresh start in the next update.

For these two features, if you really need to use them (I'm probably talking to myself here, lol), the most
I can say is to look into the source for this legacy build, and how STE used it also. It's not even using it much
there, so that's probably why I didn't smoothen it out yet. I think the technical debt for this component got
pretty high for me with the last few updates too, so that didn't help with being able to add new things safely.
Hence, why this whole rework is happening.
```