Dist Type Checking (Oops)
Looks like with the dual-config approach, it was type checking the `./dist` folder too, ouch! This means NBTify is doing this too, and I just hadn't noticed it, because none of the exports/definitions collide with those in the rest of the project. It looks to be because of Menu-Drop's `declare global` statements for the `HTMLElementTagNameMap`, which makes sense.

Glad I caught this! Setting this up has been a bit complex, both in general, and for NBTify, where I first had to figure things out. It's that I want full type-checking across the project's files, but I don't want to *include* all of those files in the build output. For NBTify, that includes the `./test` folder, with it's demos and it's full test runner. Here, I will want it for the test folder's demos eventually, and the source folder. I can also just use `// @ts-check` and write in JS for the time being, for the test demo page. The other option is to set it up with Vite maybe, and write full TS in the demo too. That is probably what I'll eventually look into, in regards with better managing ES-Module-Shims and Constructable Stylesheets polyfills.