AllowArbitraryExtensions
This feels like a more secure/explicit way to import Constructable Stylesheets, compared to `declare module "*.css"`, because TSC won't validate that the file you are importing exists, which I'm not a complete fan of. But this also means that each stylesheet file will need it's own type declaration, which I also don't like haha. So, we'll see.

https://www.typescriptlang.org/tsconfig#allowArbitraryExtensions

Also noticed I forgot to include the updated `.npmignore` for the dual-config TSC setup.

Flamingosis
https://www.youtube.com/watch?v=e6-HossxeWc

Flying Beagle
https://www.youtube.com/watch?v=HHOn8u-c2wk