Constructable Stylesheet Types
Realized that the types being in the `./styles/` directory is actually ok, because that ensures that the CSS file gets the proper types it needs.

For the event I decide to do this down the line, I wanted to ensure that the library's internal types for the CSS file wouldn't leak out and cause unexpected errors. Here's the example of that which I mean:

```ts
// ./src/index.ts
export { default as stylesheet } from "../styles/style.css" assert { type: "css" };
```

I thought this was an issue with AllowArbitraryExtensions, so I tried going back to `declare module "*.css"`. That had the same error though, in that neither file was included in the output TS, regardless of whether the definition file was in `./src/`, or `./styles`. I thought that was what was causing the missing types, but it was more because `.d.ts` files written by the user aren't kept around in the build output by TS. This makes sense in the same way that `global.d.ts` doesn't get exported either, it's the mirroring of that behavior. So, the way to keep things working is to allow the `./styles/style.d.css.ts` file to be kept as part of the build output, as it is necessary for cases that you export the stylesheet from the main library's code, which I may want to do/want to make sure doesn't go un-typed accidentally if I do randomly export it from another module.

Here's a second reference to this just in-case.
https://www.typescriptlang.org/tsconfig#allowArbitraryExtensions