MenuItemLike (Demo)
I think I thought of doing something like this a few months ago. Having tried it out though, I feel like It doesn't quite completely make sense for this case though, so I'll just commit it for the demo. If the overlap in properties was more of a specific case for things, maybe it makes sense to use a dedicated type, but it's only for accessing a few properties in this case, so I think using TS's strengths in what comes with plain union types may make more sense here. I think it's a similar argument to that of explicit return types. This solution here would be the example of explicit return types, and the union-driven property access would be the implicit return type version.

Personally, I've started to like using explicit return types, for library code at least (and more-and-more, other stuff too, regular web dev and scripts), because it can help ensure that the function meets the criteria you need it to. I think having checks on cases like these, where you want to ensure the nullish representation of a value is accurate, is something you want to narrow down on. I don't want any `undefined`, or even unintended `null` values leaking around the codebase for things like this. Return types and property index access have helped out a bunch in detecting unsafe method calls and deep-property access (accessing properties on properties on properties).

This example also feels similar to that of CodeAesthetic's video on dependency injection.
https://www.youtube.com/watch?v=J1f5b4vcxCQ