CSS Import Types
This is the updated syntax for things. The other one changed as of the last year or two I think. I haven't been keeping up with the new updates in the web platform the last year or two, been focusing more on my own projects and more systems-level programming, at least for local types of programs, like in Rust and or Node.js.

Outflippingstanding.
https://maggottwat-us.bandcamp.com/track/sexy-plants