Safari Mouse PointerType Fix
Well, and attempt at it. This is just plain not working, really. This library is such a mess, as is STE. It was doing so well, but I kind of just lost upkeep with it being bug-free, as well as working on the new implementation for it. Just started getting messy, and losing interest on it. But I have a site that uses it, so I inherently need to address this sooner than later. I mean, the project(s) as a whole. I very much still love these projects. They are not what they once were, if you were to look at it that way. My lack of skills for upkeep early on really held it back all of this time later. That's bound to happen. It is unfortunate though. But again, it had to happen that way for me to learn what to do, and not to do.