PointerType Fix
Realized this was an issue for some time now. At some point, Safari added support for event.pointerType on click events, but it still reports it as mouse, even though on the preceding pointerdown event, it is still showing correctly as touch. So this disparity prevents the handler from opening correctly on Safari-based touch devices. Very annoying.

This seems to keep things working the same across desktop and mobile, so I am keeping it for the meantime. Less than ideal, but hey it seems to work. Happy to have any other better ways to handle this distinction, if you know of any.

STE has been broken like this for a while now, where the menus plain just won't open on touch. Glad to finally have debugged this.

STE really needs some more love, it is really falling behind bad. It is so close to being there, just need to rework some final few things over to SolidJS completely. Almost completely un-spaghettied the codebase! There was a lot of meat sauce to clean up, it really stained the code quality. ha